added linear and nonlinear data generating process benchmarking documentation; included mathematical model details, implementation scripts, and relevant examples
