+++ docs/research.md
@@ -1,1 +1,7 @@
 # Research
+
+```{toctree}
+:maxdepth: 2
+
+research/dgp_benchmarking
+```
